Added BuildSchema overload allowing for customizable column types.
Removed PopulateTableFromCSV overloads that do not require a hasHeaders boolean.

Added overloads for PopulateTableFromCSV to allow for full file path or separation of folder path and file name in parameters.

PopulateTableFromCSV methods all now import from file rather than a list of strings.

Many methods are now public instead of private to more appropriately fit the requirements of the library.

Added my own method to add rows to tables. Operates identically to pre-existing extensions, but allows for a different sense of readability.